requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5.1
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-filter==2.2.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==2.1.2
djangorestframework==3.11.0
easy-thumbnails==2.7
entrypoints==0.3
flake8==3.7.9
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
peewee==3.13.2
Pillow==7.1.1
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.4
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.8
wrapt==1.11.2
